requirements.txt
Django==2.2.3
django-guardian==2.1.0
djangorestframework==3.10.2
Markdown==3.1.1
pytz==2019.1
soupsieve==1.9.4
sqlparse==0.3.0
